2dCostEstimation
This is a small project to showcase cutting cost estimation for a 2D
part.  It is assumed that the 2D part consists of closed set of straight
lines and circular arcs.  It involves finding a minimum area best fit
rectangular blank as material cost.  And the cutting cost based on total
length to be cut. Cutting cost is assumed to be linear with length for
straight cuts and a function of radius for circular arc cuts.